+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### **(V.7.0.25 Changes) (1.20.1 Minecraft)**
+
+##### Mod Compat:
+(Forge): Checks tool actions now along with an instanceof check. 
+ This means Bumblezone's carvable blocks can be carved with more modded shears.
+ Comb Cutter can be put on more modded shears/swords too.
+ And more modded axes can disable Honey Crystal Shield.
+
+
 ### **(V.7.0.24 Changes) (1.20.1 Minecraft)**
 
 ##### Items: